+++ src/js/lotto/controller.js
@@ -1,8 +1,8 @@
 import LottoModel from "./model.js";
-import LottoView from "./view.js";
-import Validator from "./validator/validator.js";
-import { $, getQuotient, getRandomNumber, sortByNumber } from "./util.js";
-import { LOTTO, SELECTOR } from "./constant.js";
+import LottoView from "../view.js";
+import Validator from "./validator.js";
+import { $, getQuotient, getRandomNumber, sortByNumber } from "../util.js";
+import { LOTTO, SELECTOR } from "../constant.js";
 
 class LottoController {
   constructor() {
@@ -17,7 +17,7 @@ class LottoController {
     this.handlePrice();
   }
 
-  getAndClearBuyValue() {
+  getandClearBuyInputValue() {
     const value = $(SELECTOR.BUY_INPUT).value;
     $(SELECTOR.BUY_INPUT).value = "";
