+++ src/js/lotto/validator.js
@@ -1,5 +1,17 @@
 import { ERROR_MESSAGE, LOTTO } from "../constant.js";
 class Validator {
+  isFloatPrice(price) {
+    return parseInt(price, 10) !== price;
+  }
+
+  isNegativeNumber(price) {
+    return price < 0;
+  }
+
+  isLessThanThousand(price) {
+    return 0 <= price && price < LOTTO.PRICE;
+  }
+
   isPriceValid(price) {
     if (this.isFloatPrice(price)) {
       return ERROR_MESSAGE.PRICE_CANNOT_BE_FLOAT;
@@ -14,18 +26,6 @@ class Validator {
     }
 
     return null;
-  }
-
-  isFloatPrice(price) {
-    return parseInt(price, 10) !== price;
-  }
-
-  isNegativeNumber(price) {
-    return price < 0;
-  }
-
-  isLessThanThousand(price) {
-    return 0 <= price && price < LOTTO.PRICE;
   }
 }
