+++ src/js/lotto/view.js
